:recycle: refactor(lib/pdf): Factor out isomorphic canvas helpers.
